Update state circuit spec from impl
The main change is that the lexicographic ordering now includes the RWC
as well (so we don't need to check the RWC is monotonically increasing
individually); and that the lexicographic ordering uses a LowerThan
comparison collecting all keys into 16 bit limbs.

Update the spec notes to use two numbers to be more clear.

Use column key names instead of key indexes to make the constraints more
clear (like in the implementation)