feat: Include AccountStorage existance support
In order to make account_storage reads/writes we need first to assert
that the storage for a particular account exists. Not only the account
but also the storage associated to it.

This adds into the table specs and implementation a new tag for
`AccountStorage` called `AccountStorageTag`. This allows to lookup the
existance of an account storage.

Also includes the conversion between account_storage_tag ->
MPTProofType.